+++ pydfs_lineup_optimizer/sites/fantasy_draft/settings.py
@@ -3,12 +3,14 @@ from pydfs_lineup_optimizer.settings import BaseSettings, LineupPosition
 from pydfs_lineup_optimizer.constants import Sport, Site
 from pydfs_lineup_optimizer.sites.sites_registry import SitesRegistry
 from pydfs_lineup_optimizer.lineup_printer import IndividualSportLineupPrinter
+from pydfs_lineup_optimizer.sites.fantasy_draft.importer import FantasyDraftCSVImporter
 
 
 class FantasyDraftSettings(BaseSettings):
     site = Site.FANTASY_DRAFT
     budget = 100000
     max_from_one_team = 6  # type: Optional[int]
+    csv_importer = FantasyDraftCSVImporter
 
 
 @SitesRegistry.register_settings