+++ pydfs_lineup_optimizer/sites/yahoo/settings.py
@@ -1,12 +1,14 @@
 from pydfs_lineup_optimizer.settings import BaseSettings, LineupPosition
 from pydfs_lineup_optimizer.constants import Sport, Site
 from pydfs_lineup_optimizer.sites.sites_registry import SitesRegistry
+from pydfs_lineup_optimizer.sites.yahoo.importer import YahooCSVImporter
 
 
 class YahooSettings(BaseSettings):
     site = Site.YAHOO
     budget = 200
     max_from_one_team = 6
+    csv_importer = YahooCSVImporter
 
 
 @SitesRegistry.register_settings